[PIM-6990] Fix reviews
