+++ .ci/behat.community.yml
@@ -86,6 +86,8 @@ default:
                     - 'Context\FeatureContext'
                 - Pim\Behat\Context\Storage\VariantProductStorage:
                     - '@pim_catalog.repository.product'
+                - Pim\Behat\Context\Domain\Enrich\ItemPickerContext:
+                    - 'Context\FeatureContext'
     extensions:
         Behat\ChainedStepsExtension: ~
         Behat\MinkExtension:
